Rename DataPickerContainer to DataPicker - refactor default export to named export
